Refactor Playfield sizing and introduce `PlayfieldSizing`; improve thread safety and dispose logic; add `PlayfieldContainer`

+++ ThirtyDollarVisualizer/Objects/DollarStoreCamera.cs
@@ -108,7 +108,7 @@ public sealed class DollarStoreCamera : Camera
         base.SetMatrixValue(left, right, bottom, top);
     }
 
-    public void Update(float secondsLastFrame)
+    public void Update(double secondsLastFrame)
     {
         var current = Position;
         var target = _virtualPosition;
